Populate, add reactions, query post and user, add Token to POSTS

+++ src/services/Posts/index.js
@@ -3,12 +3,13 @@ const cloudinary = require("cloudinary").v2;
 const { CloudinaryStorage } = require("multer-storage-cloudinary");
 const multer = require("multer");
 const mongoose = require("mongoose");
+const { verifyToken } = require("../../utilities/errorHandler");
+const jwt = require("jsonwebtoken");
 // const fs = require("fs");
 // const MongoClient = require("mongodb").MongoClient;
 // const url = "mongodb://localhost:5001/";
 // const Json2csvParser = require("json2csv").Parser;
 // const PDFDocument = require("pdfkit");
-const q2m = require("query-to-mongo");
 
 const Post = require("./schema");
 
@@ -51,45 +52,50 @@ const parser = multer({ storage: storage });
 //   }
 // });
 
-route.post("/:id/upload", parser.single("image"), async (req, res, next) => {
+route.post("/:id/upload", verifyToken, parser.single("image"), async (req, res, next) => {
   try {
-    // const newPost = {
-    //   ...req.body,
-    //   image: req.file.path,
-    // };
+    jwt.verify(req.token, secretKey, async (err, data) => {
+      if (err) res.sendStatus(403);
+      else {
+        if (await Post.findOne({ $and: [{ _id: req.params.id }, { user: data._id }] })) {
+          const modifiedPost = await Post.findByIdAndUpdate(
+            req.params.id,
+            {
+              $set: {
+                image: req.file.path,
+              },
+            },
+            {
+              useFindAndModify: false,
+            }
+          );
 
-    const modifiedPost = await Post.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          image: req.file.path,
-        },
-      },
-      {
-        useFindAndModify: false,
+          console.log(req.file.path);
+          res.status(200).send(modifiedPost);
+        } else res.sendStatus(403);
       }
-    );
-
-    console.log(req.file.path);
-    res.status(200).send(modifiedPost);
+    });
   } catch (error) {
     console.log(error);
     next(error);
   }
 });
 
-//----Not yet done----///
-
 //<-------------------------------------------------^ Image Upload ^---------------------------------------------------->//
-route.post("/", async (req, res, next) => {
+route.post("/", verifyToken, async (req, res, next) => {
   try {
-    const myObj = {
-      ...req.body,
-      image: "https://miro.medium.com/max/10368/1*o8tTGo3vsocTKnCUyz0wHA.jpeg",
-    };
-    const newPost = new Post(myObj);
-    await newPost.save();
-    res.status(201).send(newPost);
+    jwt.verify(req.token, secretKey, async (err, data) => {
+      if (err && req.body.user !== data._id) res.sendStatus(403);
+      else {
+        const myObj = {
+          ...req.body,
+          image: "https://miro.medium.com/max/10368/1*o8tTGo3vsocTKnCUyz0wHA.jpeg",
+        };
+        const newPost = new Post(myObj);
+        await newPost.save();
+        res.status(201).send(newPost);
+      }
+    });
   } catch (error) {
     console.log(error);
     next(error);
@@ -98,10 +104,11 @@ route.post("/", async (req, res, next) => {
 
 route.get("/", async (req, res, next) => {
   try {
-    const query = q2m(req.query);
-    console.log(query);
-    const allPost = await Post.find().sort({ createdAt: -1 });
-    res.status(200).send(allPost);
+    const allPost = await Post.find(req.query.search && { $text: { $search: req.query.search } })
+      .sort({ createdAt: -1 })
+      .populate("author", "name surname image title");
+    const posts = allPost.map((post) => (post = { ...post, reactions: reactions.length }));
+    res.status(200).send(posts);
   } catch (error) {
     console.log(error);
     next(error);
@@ -110,7 +117,7 @@ route.get("/", async (req, res, next) => {
 
 route.get("/:id", async (req, res, next) => {
   try {
-    const singlePost = await Post.findById(req.params.id);
+    const singlePost = await Post.findById(req.params.id).populate("author", "name surname image title").populate("reactions.user", "name surname image title");
     res.status(200).send(singlePost);
   } catch (error) {
     console.log(error);
@@ -118,22 +125,37 @@ route.get("/:id", async (req, res, next) => {
   }
 });
 
-route.put("/:id", async (req, res, next) => {
+route.put("/:id", verifyToken, async (req, res, next) => {
   try {
-    const modifiedPost = await Post.findByIdAndUpdate(req.params.id, req.body, {
-      useFindAndModify: false,
+    jwt.verify(req.token, secretKey, async (err, data) => {
+      if (err) res.sendStatus(403);
+      else {
+        if (await Post.findOne({ $and: [{ _id: req.params.id }, { user: data._id }] })) {
+          const modifiedPost = await Post.findByIdAndUpdate(req.params.id, req.body, {
+            useFindAndModify: false,
+          });
+          res.status(200).send(modifiedPost);
+        } else res.sendStatus(403);
+      }
     });
-    res.status(200).send(modifiedPost);
   } catch (error) {
     console.log(error);
     next(error);
   }
 });
 
-route.delete("/:id", async (req, res, next) => {
+route.delete("/:id", verifyToken, async (req, res, next) => {
   try {
-    const deletedPost = await Post.findByIdAndDelete(req.params.id);
-    res.status(200).send("POST DELETED");
+    jwt.verify(req.token, secretKey, async (err, data) => {
+      if (err) res.sendStatus(403);
+      else {
+        if (await Post.findOne({ $and: [{ _id: req.params.id }, { user: data._id }] })) {
+          await Post.findByIdAndDelete(req.params.id);
+          res.status(200).send("POST DELETED");
+        }
+        res.sendStatus(403);
+      }
+    });
   } catch (error) {
     console.log(error);
     next(error);
